tests/web_platform_tests/graph/shared-graph-sync.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>SharedGraph P2P Sync — WPT</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';

// ==========================================================================
// Spec 03 — Peer-to-Peer Graph Synchronisation Protocol
// https://github.com/HexaField/w3c-living-web-proposals/blob/main/drafts/03_p2p-graph-sync.md
// ==========================================================================

// §5.1 SharedGraphManager — share() and join()

promise_test(async t => {
  assert_true(typeof navigator.graph.join === 'function', 'join() MUST exist');
  assert_true(typeof navigator.graph.listShared === 'function', 'listShared() MUST exist');
}, '§5.1 PersonalGraphManager MUST expose join() and listShared()');

promise_test(async t => {
  const graph = await navigator.graph.create('Shareable');
  assert_true(typeof graph.share === 'function', 'share() MUST exist on PersonalGraph');
  await navigator.graph.remove(graph.uuid);
}, '§5.1 PersonalGraph MUST expose share()');

promise_test(async t => {
  const graph = await navigator.graph.create('ShareTest');
  const shared = await graph.share({
    meta: { name: 'Test Community' }
  });
  assert_true(shared instanceof SharedGraph, 'MUST return SharedGraph');
  assert_true(shared.uri.length > 0, 'MUST have a URI');
  await navigator.graph.remove(graph.uuid);
}, '§8.1 share() MUST return a SharedGraph with a globally unique URI');

// §5.2 SharedGraph interface

promise_test(async t => {
  const graph = await navigator.graph.create('SharedIface');
  const shared = await graph.share();

  // SharedGraph extends PersonalGraph
  assert_true(shared instanceof PersonalGraph, 'MUST extend PersonalGraph');
  assert_true(shared instanceof EventTarget, 'MUST extend EventTarget');

  // Attributes
  assert_true('uri' in shared, 'MUST have uri');
  assert_true('syncState' in shared, 'MUST have syncState');

  // SyncState enum values
  const validStates = ['idle', 'syncing', 'synced', 'error'];
  assert_true(validStates.includes(shared.syncState), 'syncState MUST be valid enum');

  await navigator.graph.remove(graph.uuid);
}, '§5.2 SharedGraph MUST have uri and syncState attributes');

promise_test(async t => {
  const graph = await navigator.graph.create('PeerOps');
  const shared = await graph.share();

  assert_true(typeof shared.peers === 'function', 'peers() MUST exist');
  assert_true(typeof shared.onlinePeers === 'function', 'onlinePeers() MUST exist');

  const peers = await shared.peers();
  assert_true(Array.isArray(peers), 'peers() MUST return array');

  const online = await shared.onlinePeers();
  assert_true(Array.isArray(online), 'onlinePeers() MUST return array');

  await navigator.graph.remove(graph.uuid);
}, '§5.2 SharedGraph MUST expose peers() and onlinePeers()');

// §9 Signalling

promise_test(async t => {
  const graph = await navigator.graph.create('Signals');
  const shared = await graph.share();

  assert_true(typeof shared.sendSignal === 'function', 'sendSignal MUST exist');
  assert_true(typeof shared.broadcast === 'function', 'broadcast MUST exist');

  // broadcast should not throw even with no peers
  await shared.broadcast({ type: 'ping' });

  await navigator.graph.remove(graph.uuid);
}, '§9.1-9.2 SharedGraph MUST expose sendSignal() and broadcast()');

// §8.2 Joining

promise_test(async t => {
  const graph = await navigator.graph.create('Joinable');
  const shared = await graph.share({ meta: { name: 'Join Test' } });

  const joined = await navigator.graph.join(shared.uri);
  assert_true(joined instanceof SharedGraph, 'MUST return SharedGraph');
  assert_equals(joined.uri, shared.uri, 'MUST have same URI');

  await navigator.graph.remove(graph.uuid);
}, '§8.2 join() MUST connect to existing SharedGraph by URI');

// §5.5 SyncState enum

test(t => {
  // SyncState values are tested indirectly through SharedGraph.syncState above.
  // This test verifies the idle default.
}, '§5.5 SyncState MUST include idle, syncing, synced, error');

// SharedGraph inherits PersonalGraph operations

promise_test(async t => {
  const graph = await navigator.graph.create('SharedTriples');
  const shared = await graph.share();

  // All PersonalGraph methods must be available
  assert_true(typeof shared.addTriple === 'function');
  assert_true(typeof shared.queryTriples === 'function');
  assert_true(typeof shared.snapshot === 'function');
  assert_true(typeof shared.addShape === 'function');

  await navigator.graph.remove(graph.uuid);
}, '§4.1 SharedGraph MUST support all PersonalGraph operations');
</script>
